refactor: improve next and typescript for project
refactor: rename from search to SearchLayout

chore: remove unused imports

refactor: use document for links and favicon

fix: fix typescript type issue
